eliminate compiler warnings

+++ MagickCore/image.c
@@ -908,18 +908,20 @@ MagickExport Image *CloneImage(const Image *image,const size_t columns,
     scale_x=(double) columns/(double) image->columns;
   if (image->rows != 0)
     scale_y=(double) rows/(double) image->rows;
-  clone_image->page.width=(size_t) CastDoubleToSsizeT(floor(scale_x*
-    image->page.width+0.5));
-  clone_image->page.height=(size_t) CastDoubleToSsizeT(floor(scale_y*
-    image->page.height+0.5));
+  clone_image->page.width=(size_t) CastDoubleToSsizeT(floor((double) scale_x*
+    (double) image->page.width+0.5));
+  clone_image->page.height=(size_t) CastDoubleToSsizeT(floor((double) scale_y*
+    (double) image->page.height+0.5));
   if (MagickAbsoluteValue(scale_x-scale_y) < 2.0)
     scale_x=scale_y=MagickMin(scale_x,scale_y);
-  clone_image->page.x=CastDoubleToSsizeT(ceil(scale_x*image->page.x-0.5));
-  clone_image->tile_offset.x=CastDoubleToSsizeT(ceil(scale_x*
-    image->tile_offset.x-0.5));
-  clone_image->page.y=CastDoubleToSsizeT(ceil(scale_y*image->page.y-0.5));
-  clone_image->tile_offset.y=CastDoubleToSsizeT(ceil(scale_y*
-    image->tile_offset.y-0.5));
+  clone_image->page.x=CastDoubleToSsizeT(ceil((double) scale_x*
+    (double) image->page.x-0.5));
+  clone_image->tile_offset.x=CastDoubleToSsizeT(ceil((double) scale_x*
+    (double) image->tile_offset.x-0.5));
+  clone_image->page.y=CastDoubleToSsizeT(ceil((double) scale_y*(double)
+    (double) image->page.y-0.5));
+  clone_image->tile_offset.y=CastDoubleToSsizeT(ceil((double) scale_y*
+    (double) image->tile_offset.y-0.5));
   clone_image->cache=ClonePixelCache(image->cache);
   if (SetImageExtent(clone_image,columns,rows,exception) == MagickFalse)
     clone_image=DestroyImage(clone_image);
@@ -1763,7 +1765,7 @@ MagickExport size_t InterpretImageFilename(const ImageInfo *image_info,
             pattern_length=strlen(pattern);
             if ((count <= 0) || ((size_t) count != pattern_length))
               return(0);
-            if ((p-filename+pattern_length) >= MagickPathExtent)
+            if ((p-filename+(ssize_t) pattern_length) >= MagickPathExtent)
               return(0);
             (void) CopyMagickString(p,pattern,(size_t) (MagickPathExtent-
               (p-filename)));
@@ -1810,7 +1812,7 @@ MagickExport size_t InterpretImageFilename(const ImageInfo *image_info,
         if (option == (const char *) NULL)
           continue;
         option_length=strlen(option);
-        if ((p-filename+option_length) >= MagickPathExtent)
+        if ((p-filename+(ssize_t) option_length) >= MagickPathExtent)
           return(0);
         (void) CopyMagickString(p,option,(size_t) (MagickPathExtent-
           (p-filename)));
